+++ src/Backend/Modules/Mailmotor/DependencyInjection/MailmotorExtension.php
@@ -9,10 +9,10 @@ namespace Backend\Modules\Mailmotor\DependencyInjection;
  * file that was distributed with this source code.
  */
 
+use Backend\Modules\Mailmotor\DependencyInjection\Compiler\MailmotorCompilerPass;
 use Symfony\Component\Config\FileLocator;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\HttpKernel\DependencyInjection\Extension;
-use Symfony\Component\DependencyInjection\Compiler\CompilerPassInterface;
 use Symfony\Component\DependencyInjection\Loader;
 
 /**
@@ -20,7 +20,7 @@ use Symfony\Component\DependencyInjection\Loader;
  *
  * To learn more see {@link http://symfony.com/doc/current/cookbook/bundles/extension.html}
  */
-class MailmotorExtension extends Extension implements CompilerPassInterface
+class MailmotorExtension extends Extension
 {
     /**
      * {@inheritdoc}
@@ -29,35 +29,7 @@ class MailmotorExtension extends Extension implements CompilerPassInterface
     {
         $loader = new Loader\YamlFileLoader($container, new FileLocator(__DIR__ . '/../Resources/config'));
         $loader->load('services.yml');
-    }
 
-    /**
-     * {@inheritdoc}
-     */
-    public function process(ContainerBuilder $container)
-    {
-        // we have the service fork.settings and it's not empty
-        if ($container->has('fork.settings') && !is_a($container->get('fork.settings'), 'stdClass')) {
-            // we must set these parameters to be usable
-            $container->setParameter(
-                'mailmotor.mail_engine',
-                $container->get('fork.settings')->get('Mailmotor', 'mail_engine')
-            );
-            $container->setParameter(
-                'mailmotor.api_key',
-                $container->get('fork.settings')->get('Mailmotor', 'api_key')
-            );
-            $container->setParameter(
-                'mailmotor.list_id',
-                $container->get('fork.settings')->get('Mailmotor', 'list_id')
-            );
-        // when in fork cms installer, we don't have the service fork.settings
-        // but we must set the parameters
-        } else {
-            // we must set these parameters to be usable
-            $container->setParameter('mailmotor.mail_engine', 'not_implemented');
-            $container->setParameter('mailmotor.api_key', null);
-            $container->setParameter('mailmotor.list_id', null);
-        }
+        $container->addCompilerPass(new MailmotorCompilerPass());
     }
 }
